+++ test/ResourceAdequacy/results/debug.jl
@@ -26,7 +26,7 @@
     gens_available = lines_available = stors_available = genstors_available =
         zeros(Bool, 0, 0, 0)
 
-    sample_ues = zeros(Int, 0)
+    sample_ues = zeros(Float64, 0)
 
     result = ResourceAdequacy.DebugResult(
         regions, interfaces, tstamps,
